requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
bleach==6.0.0
Django==5.0.1
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pytz==2023.3
six==1.16.0
webencodings==0.5.1
django-nose==1.4.7
nose==1.3.7
pynose==1.4.8
lorem_text
django-prometheus
pynose==1.4.8
pillow==10.2.0
pinocchio==0.4.3
nose==1.3.7
setuptools
coverage==7.4.1
psycopg2-binary==2.9.9
